Feature:  Player can use the mouse to move as well now
**Important Note**:
Had to embed the canvas element directly into the html file.
Previously, the engine created the canvas.  But since the engine loads
after the app.js file, there wasn't a nice way to setup an
event listener on the canvas.  Now that the canvas already exists,
it's good to go.