Fix:  Hit detection now works correctly for mouse input
The refactorization didn't cause the error.  It was because
I didn't update the mouse input logic with the new requirements
for hit detection.

So... even though I was trying to save on unnecessary calls
during update(), had I just put it in there (which DOES make sense),
this bug could've been avoided.  So... which is more important?
Readability or efficiency?